% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/watersheds.r
\name{create_watershed}
\alias{create_watershed}
\title{Create a watershed in the database}
\usage{
create_watershed(conn, name, schema = "watersheds")
}
\arguments{
\item{conn}{A \code{\link[=fleeDB_connection]{fleeDB_connection()}} with write permissions}

\item{name}{The watershed name (e.g., "vjosa")}

\item{schema}{Where to write it, normally do not change the default}
}
\value{
NULL, has the side effect of creating watershed tables in the database
}
\description{
Create a watershed in the database
}
