% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/flee_db.r
\name{fdb_connect}
\alias{fdb_connect}
\title{Connect to the fleeDB}
\usage{
fdb_connect(key_name = "fleedb-login-c7701212", user = "c7701212")
}
\arguments{
\item{key_name}{The name of the fleeDB key stored on your computer}
}
\value{
A db connection from the RPostgres package
}
\description{
Connect to the fleeDB
}
