% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/watersheds.r
\name{fdb_streams}
\alias{fdb_streams}
\title{Read stream map from the database}
\usage{
fdb_streams(conn, name, schema = "watersheds")
}
\arguments{
\item{conn}{A \code{\link[=fleeDB_connection]{fleeDB_connection()}} with write permissions}

\item{name}{The watershed name (e.g., "vjosa")}

\item{schema}{Where to write it, normally do not change the default}
}
\value{
A map of streams in sf format
}
\description{
Read stream map from the database
}
