% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/watersheds.r
\name{write_watershed}
\alias{write_watershed}
\title{Writes watershed data to the database}
\usage{
write_watershed(
  conn,
  name,
  stream,
  pixels,
  pix_time,
  Tp,
  Tr,
  schema = "watersheds"
)
}
\arguments{
\item{conn}{A \code{\link[=fleeDB_connection]{fleeDB_connection()}} with write permissions}

\item{name}{The watershed name (e.g., "vjosa")}

\item{stream}{A vector map of streams, in sf format}

\item{pixels}{A data.frame of pixel data, with IDs}

\item{pix_time}{A data.frame of pixel-time data (e.g., discharge etc)}

\item{Tp}{A pixel topology, in SparseMatrix format}

\item{Tr}{A reach topology SparseMatrix}

\item{schema}{Where to write it, normally do not change the default}
}
\value{
NULL, has the side effect of writing watershed tables in the database
}
\description{
Writes watershed data to the database
}
\details{
Column names in the tables MUST match those in the database.
}
